feat(resource creation): integrate upload form
