+++ src/app/workspace/resource/resource-instance-form/select-properties/select-properties.component.html
@@ -1,5 +1,10 @@
 <div class="properties-container">
     <div class="properties">
+
+        <div *ngIf="hasFileValue?.length > 0">
+            <dsp-upload-file #upload [representation]="hasFileValue[0].objectType" [parentForm]="parentForm"></dsp-upload-file>
+        </div>
+
         <div *ngFor="let prop of properties; let last = last;" [class.border-bottom]="!last">
             <div class="property" *ngIf="!prop.isLinkProperty || prop">
                 <div class="property-label">
